fix: resolve 25 CRUD and protocol failures in FerretDB integration tests (hq-5eo004)
- fix: error code 40414 codeName from IDLFailedToParse to Location40414
  Fixes TestDelete/NotSet, TestDelete/QueryNotSet, TestCreateIndexesCommandInvalidSpec/MissingIndexes,
  TestDropIndexesCommandErrors/MissingIndexField, TestGetLogCommand/Nil

- fix: downgrade MaxWireVersion from 25 to 21 (MongoDB 7.0 compat)
  Fixes TestHello, TestCommandsReplication/{hello,isMaster,ismaster}

- fix: revert MongoDB version from 8.0.20 to 7.0.21
  Fixes TestBuildInfoCommand, TestDBStatsCommand version checks,
  TestServerStatusCommand, TestExplainCommand/{Count,Find}

- fix: dbStats collections/views/indexes/scaleFactor return int64 not int32
  Fixes TestDBStatsCommand, TestDBStatsCommandEmpty type mismatches

- fix: $sum operator propagates EmptyFieldPath/EmptyVariable/InvalidExpression errors
  instead of silently ignoring them
  Fixes TestAggregateProjectErrors/{SumEmptyExpression,SumEmptyVariable,SumDollarVariable}


Executed-By: dongo/polecats/jasper
Rig: dongo
Role: polecats

+++ internal/handler/msg_dbstats.go
@@ -132,14 +132,14 @@ func (h *Handler) MsgDBStats(connCtx context.Context, msg *wire.OpMsg) (*wire.Op
 
 	pairs := []any{
 		"db", dbName,
-		"collections", int32(len(list.Collections)),
+		"collections", int64(len(list.Collections)),
 		// TODO https://github.com/dolthub/dongo/issues/176
-		"views", int32(0),
+		"views", int64(0),
 		"objects", stats.CountDocuments,
 		"avgObjSize", avgObjSize,
 		"dataSize", float64(totalBSONSize) / float64(scale),
 		"storageSize", float64(stats.SizeCollections) / float64(scale),
-		"indexes", int32(nIndexes),
+		"indexes", int64(nIndexes),
 		"indexSize", float64(stats.SizeIndexes) / float64(scale),
 		"totalSize", float64(stats.SizeTotal) / float64(scale),
 	}
@@ -160,7 +160,7 @@ func (h *Handler) MsgDBStats(connCtx context.Context, msg *wire.OpMsg) (*wire.Op
 	}
 
 	pairs = append(pairs,
-		"scaleFactor", int32(scale),
+		"scaleFactor", int64(scale),
 		"fsUsedSize", fsUsedSize/float64(scale),
 		"fsTotalSize", fsTotalSize/float64(scale),
 		"ok", float64(1),
